main.py: use the USSD code as id
Until now, one had to figure out which package was which by comparing its
price, volume, and duration.

With this commit, we also read the ussd-code from the website which acts as
an identification. Also, one could actually dial the USSD code to buy the
package!

Notice the trailing white-space in the output file. This white-space is
used in other programs (i.e. gnuastro table) for calculations and/or to
show the metadata. So be careful with your text editor!

Mohammad suggested using the 'USSD' as the identification.